+++ src/main/java/com/frejdh/util/job/model/callables/JobOnError.java
@@ -1,6 +1,8 @@
 package com.frejdh.util.job.model.callables;
 
+import com.frejdh.util.job.Job;
+
 @FunctionalInterface
 public interface JobOnError {
-	void onError(Throwable throwable);
+	void onError(Job job, Throwable throwable);
 }