+++ src/main/java/com/frejdh/util/job/model/callables/JobOnFinalize.java
@@ -1,6 +1,8 @@
 package com.frejdh.util.job.model.callables;
 
+import com.frejdh.util.job.Job;
+
 @FunctionalInterface
 public interface JobOnFinalize {
-	void onComplete();
+	void onComplete(Job job);
 }