+++ src/main/java/com/frejdh/util/job/model/callables/JobOnStatusChange.java
@@ -1,6 +1,8 @@
 package com.frejdh.util.job.model.callables;
 
+import com.frejdh.util.job.Job;
+
 @FunctionalInterface
 public interface JobOnStatusChange {
-	void onStatusChange();
+	void onStatusChange(Job job);
 }